requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
async-timeout==4.0.2
asynctest==0.13.0
attrdict==2.0.1
attrs==23.1.0
Babel==2.12.1
bce-python-sdk==0.8.83
bcrypt==4.0.1
beautifulsoup4==4.12.2
bleach==6.0.0
Brotli==1.0.9
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.34
defusedxml==0.7.1
dgl==1.1.0
einops==0.2.0
et-xmlfile==1.1.0
fastapi==0.95.2
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.12.0
fire==0.5.0
Flask==2.2.5
flask-babel==3.1.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gdown==4.4.0
gevent==22.10.2
geventhttpclient==2.0.2
gradio==3.11.0
greenlet==2.0.2
grpcio==1.54.2
h11==0.12.0
httpcore==0.15.0
httpx==0.24.0
idna==3.4
imageio==2.28.1
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
ipython-genutils==0.2.0
ipywidgets==8.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
#jupyter==1.0.0
#jupyter-console==6.6.3
#jupyter-server==1.24.0
#jupyter_core==4.12.0
#jupyterlab-pygments==0.2.2
#jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
linkify-it-py==2.0.2
lmdb==1.4.1
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.5.3
mdit-py-plugins==0.3.5
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
networkx==2.6.3
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.6
ocrd-fork-pylsd==0.0.3
#onnx==1.14.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.8.12
paddle-bfloat==0.1.7
paddle2onnx==1.0.6
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==1.1.5
pandocfilters==1.5.0
paramiko==3.1.0
pdf2docx==0.5.6
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
prefetch-generator==1.0.1
premailer==3.10.0
prometheus-client==0.16.0
#protobuf==3.20.2
psutil==5.9.5
pyclipper==1.3.0.post4
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.7
pydub==0.25.1
PyMuPDF==1.20.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-docx==0.8.11
python-multipart==0.0.6
python-rapidjson==1.10
pytz==2023.3
PyWavelets==1.3.0
#pywin32
#pywinpty==2.0.10
PyYAML==6.0
qtconsole==5.4.3
QtPy==2.3.1
rapidfuzz==3.0.0
rarfile==4.0
requests==2.30.0
scikit-image==0.19.3
scipy==1.7.3
Send2Trash==1.8.2
shapely==2.0.1
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
sympy==1.10.1
termcolor==2.3.0
terminado==0.17.1
tifffile==2021.11.2
tinycss2==1.2.1
torch==1.9.1
torchvision==0.10.1
tqdm==4.65.0
tritonclient==2.33.0
typing_extensions==4.5.0
uc-micro-py==1.0.2
Unidecode==1.3.6
urllib3==1.26.6
uvicorn==0.22.0
vietocr==0.3.11
visualdl==2.5.2
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wincertstore==0.2
x2paddle==1.4.1
yarl==1.9.2
zipp==3.15.0
zope.event==4.6
zope.interface==6.0
onnxruntime==1.9.0
gunicorn